Optimise SSE2 bailout.
This avoids data transfer penalties by using the float movmsk, instead
of the integer one, since the comparison was done on floats. It also
avoids an unnecessary extra comparison, and frees up a register (no need
to store zero anywhere any more).

The end of inner loop changes from

    pcmpeqb	%xmm7, %xmm1
    andps	%xmm6, %xmm15
    pmovmskb	%xmm1, %eax
    cmpl	$65535, %eax

to

    movmskps	%xmm1, %eax
    andps	%xmm6, %xmm14
    testl	%eax, %eax

which is the only major change.

The code is about 1% faster: generating an 4000x2000 image and piping it
to /dev/null goes from 0.400s to 0.396s on my CPU, as measured with `perf stat -r
3` on Linux (with openmp disabled).